requirements.txt
django>=3.2,<4.0
psycopg2-binary>=2.8
python-dateutil
stripe
openai
djangorestframework